Cleaned up and fixed Makefile
This closes #3. The export from LyX to LaTeX was improperly done
as plain latex instead of pdflatex, which is why the "EPS" files
were showing up. I also cleaned up the Makefile to make it a bit
more friendly, as well as automatically rerunning pdflatex so
that all references are updated properly.