Start of Lift 2.0 work (attribute handling)
Closes #5

Made a minor fix and decided to go ahead and start working
on the Lift 2.0 changes. The first change is a reorg and
revision of how Lift handles attributes on elements. Moved all
relevant sections to chapter 3 (fundamentals), revised attribute
binding to talk about the "-%>" operator, and made a subsection
in the forms chapter about using the varargs params on SHtml.*
to simplify attribute handling.