Major revision: moved snippets to their own chapter
Closes #13

In the process of adding details on snippet dispatch,
decided to go on a bender and move whole sections around.

It was getting to the point where it's so much info it
really deserved it. I'm probably going to do more
sectioning out in the coming weeks since the fundamentals
and advanced chapter might do better combined on a per-topic
basis. Also made a lot of little changes all over where needed.